Fix viewing drafts bug w/ nothing is published in current language
The line of code updated introduced a bug where it was not possible for a translator to view drafts in a language with nothing published.

The intention was to prevent a user who was in preview mode w/ nothing published from viewing a blank home page when leaving translator mode.  The trick of course is to not switch the language when the user is actually in translator mode.