Enable scrolling on the table view
move logic for UIRefreshControl from viewDidLoad to viewWillAppear

It will change each time the view is rendered b/c the user could enter and leave translator mode.. it's not just when the view is loaded

Remove outlet for when the refresh button was dragged

The table is now dragged, not the button.

The notification listener that was removed was unused, according to xCode

Change refresh button to just an image.