Add and remove observer for when drafts have been downloaded
When in Preview mode, after downloading a draft, there was no observer to see when the download was finished.  The resulted in the download indicator's never stopping and no further downloads were possible.

I added and removed the observer in viewDidAppear and viewDidDisappear respectively as to not interfere with observers that might be alive on GTHomeViewController