Always download the latest resources regardless of language.
Before this commit, it would only download the latest if the device locale was something other than English.  That meant English had to be updated to get the latest English version.

The bundled English should only be used as a safeguard in case the device cannot download anything on the initial open.