GT-142 Include English packages for other languages
When a language only has one or two packages translated such as Knowing
God Personally or The Four Spiritual Laws, currently the iOS version
only shows those packages. The request was to show the remaining
material in English since it is a universal language. This saves a step
for the user so they don’t have to go change the language to English to
see there is material in English.

In SetData I put code for non translator mode to get the packages from
the English language and compare the total to the current language. If
the package count is less than the English set then it will loop
through and add the English if it doesn’t exist by comparison of code.

I put code in the cellForRowAtIndexPath to check if the current
language is not english and if the package is English to “gray” it out
by changing the alpha.