Podfile and download fix
I kept running into the issue where I selected a language that had not
been downloaded yet and GodTools didn’t show any language packages in
that language since they weren’t downloaded. I didn’t click the
“Download” text but rather the row itself to the left side. Ann said
this was an issue with android that was fixed and didn’t know if it had
been fixed in iOS (apparently not since I kept finding it) I added this
piece of code to make it so the actual language clicked would be
referenced to see if it would be downloaded. This is working well and
I’m not interfering with the currentlySelected variable since that gets
set in the downloadLanguageAtCell method. I don’t think it should be
the one checked if it was downloaded because it only gets set in a few
places. Like, if I select the record to the left (checkmark) and then
hit the download text this code would work, otherwise selecting the
record tapping the left does not trigger this code since it isn’t
selected fully yet.