Making notification not show up if instructions shown
If the user originally downloads GodTools and their language is not
English, the alert would popup if that language had less materials
compared to English. This would overlay ontop of the instruction view
which would make the instructions pointless for a new user since they
would be unable to see.

I added in the dismiss selector for the instructions to set the boolean
for shouldShowInstructions so that the code further down in SetData can
check the boolean.

This would make it so the users native language would not get the
notification but they downloaded another non english language they
would see the message.