Put alert into method
I put the alert with the Untranslated package message into a method
since two instances need to call it:
It needs to display after the instruction view has shown for users who
first download the application
It also needs to display for users who already own the application and
have a language other than english. This is called from SetData

I made a persistent english Articles array and let that be referenced
in SetData as well as the dismiss code after the instruction view is
dismissed.

This means the alert will show for a user who first downloads the app
(seeing the instruction screen)
And for those who own the app already and have a main language or
selected language with less materials than English

It will only do this once per the life of the application.