iOS 10 bug: fix broken languages menu
This view was broken b/c the current iOS 10 isn't honoring background color transparencies as we're writing them now. It is honoring them when set in interface builder though. The first change therefore was to add a UIView in the language cell xib which fills the whole space and has the white background w/ 10% alpha.

This change revealed another quirk in that the accessory view wasn't filled by the view added above. The result was that the background of the accessory view was totally transparent. I decided to remove the accessory view and use straight buttons that are either shown or hidden.

Then I got to thinking that this view can be better, so I made some UI changes that feel more iOS 9.0+ including:
 - removing "Update" and "Download" words in favor of a download icon
 - moving the activity indicator and selected indicator to the right of the view where the download indicator lives.
 - reduced the font weight to  normal from bold
 - reduced the weight of the checkmark to normal or not filled
 - longer language names will be scaled down to 80% instead of truncated

The UI is better in my opinion because:

 - The languages list is now more left aligned in the cell which looks much better
 - The bold weight feels out of place in iOS 9.0+
 - Longer language names are much less likely to be truncated

I also cleaned up the interface into the LanguageViewCell class by removing "public" access to its member variables, and I stopped passing the target, selector and parameter values b/c the accessory view isn't being used any more.