Remove "Update all" button and logic
This feature makes less sense if we are removing the "Download" and "Update" wording. The app does or will soon automatically update the latest of the selected language on app reopen. Other languages that have an update will be updated the next time selected while connected to the internets.

The download button will not be shown for languages that are downloaded and have updates while not connected to the internets.